+++ packages/polaris-core/src/utils/snapshot-connectionless-util.ts
@@ -10,7 +10,7 @@ export async function getSnapshotPageById(
     if (config.connectionlessConfiguration) {
         return config.connectionlessConfiguration.getSnapshotPageById(snapshotPageId);
     } else {
-        return connection?.getRepository(SnapshotPage).findOne({} as any, snapshotPageId);
+        return connection?.getRepository(SnapshotPage).findOne(snapshotPageId);
     }
 }
 
@@ -23,7 +23,7 @@ export async function getSnapshotMetadataById(
     if (config.connectionlessConfiguration) {
         return config.connectionlessConfiguration.getSnapshotMetadataById(snapshotMetadataId);
     } else {
-        return connection?.getRepository(SnapshotMetadata).findOne({} as any, snapshotMetadataId);
+        return connection?.getRepository(SnapshotMetadata).findOne(snapshotMetadataId);
     }
 }
 
@@ -42,9 +42,7 @@ export async function saveSnapshotMetadata(
     if (config.connectionlessConfiguration) {
         return config.connectionlessConfiguration.saveSnapshotMetadata(snapshotMetadata);
     } else {
-        const metadata = await connection
-            ?.getRepository(SnapshotMetadata)
-            .save({} as any, snapshotMetadata);
+        const metadata = await connection?.getRepository(SnapshotMetadata).save(snapshotMetadata);
         return metadata instanceof Array ? metadata[0] : metadata;
     }
 }
@@ -57,7 +55,7 @@ export async function saveSnapshotPages(
     if (config.connectionlessConfiguration) {
         await config.connectionlessConfiguration.saveSnapshotPages(snapshotPages);
     } else {
-        await connection?.getRepository(SnapshotPage).save({} as any, snapshotPages);
+        await connection?.getRepository(SnapshotPage).save(snapshotPages);
     }
 }
 
@@ -73,9 +71,7 @@ export async function updateSnapshotPage(
             snapshotPageToUpdate,
         );
     } else {
-        await connection
-            ?.getRepository(SnapshotPage)
-            .update({} as any, snapshotPageId, snapshotPageToUpdate);
+        await connection?.getRepository(SnapshotPage).update(snapshotPageId, snapshotPageToUpdate);
     }
 }
 
@@ -93,7 +89,7 @@ export async function updateSnapshotMetadata(
     } else {
         await connection
             ?.getRepository(SnapshotMetadata)
-            .update({} as any, snapshotMetadataId, snapshotMetadataToUpdate);
+            .update(snapshotMetadataId, snapshotMetadataToUpdate);
     }
 }