+++ packages/polaris-middlewares/src/dates-filter/dates-filter-middleware.ts
@@ -1,9 +1,13 @@
-import {PolarisConnectionManager} from "@enigmatis/polaris-typeorm";
-import {EntityFilter, PolarisGraphQLContext, RealitiesHolder} from "@enigmatis/polaris-common";
-import {PolarisGraphQLLogger} from "@enigmatis/polaris-graphql-logger";
-import {entityFilterInputTypeName} from "@enigmatis/polaris-schema";
-import {GraphQLResolveInfo} from "graphql";
-import {isMutation} from "..";
+import { PolarisConnectionManager } from '@enigmatis/polaris-typeorm';
+import {
+    EntityFilter,
+    PolarisGraphQLContext,
+    isMutation,
+    RealitiesHolder,
+} from '@enigmatis/polaris-common';
+import { PolarisGraphQLLogger } from '@enigmatis/polaris-graphql-logger';
+import { entityFilterInputTypeName } from '@enigmatis/polaris-schema';
+import { GraphQLResolveInfo } from 'graphql';
 
 export class DatesFilterMiddleware {
     public readonly connectionManager?: PolarisConnectionManager;
@@ -28,12 +32,14 @@ export class DatesFilterMiddleware {
             context: PolarisGraphQLContext,
             info: GraphQLResolveInfo,
         ) => {
-            if(!isMutation(context.request.query)) {
+            if (!isMutation(context.request.query)) {
                 this.logger.debug('Dates filter middleware started job', context);
-                if(this.isCurrentFieldASchemaQuery(info)) {
+                if (this.isCurrentFieldASchemaQuery(info)) {
                     const datesFilterArgumentName = this.getDateFilterArgumentNameIfExists(info);
-                    if(datesFilterArgumentName) {
-                        context.entityDateRangeFilter = this.getEntityDateRangeFilter(args[datesFilterArgumentName]);
+                    if (datesFilterArgumentName) {
+                        context.entityDateRangeFilter = this.getEntityDateRangeFilter(
+                            args[datesFilterArgumentName],
+                        );
                     } else {
                         delete context.entityDateRangeFilter;
                     }
@@ -45,14 +51,14 @@ export class DatesFilterMiddleware {
                 const result = await resolve(root, args, context, info);
                 return result;
             }
-        }
+        };
     }
 
     private getDateFilterArgumentNameIfExists(info: any): string | undefined {
         let argName;
         const currentQueryArguments = info.schema.getQueryType().getFields()[info.fieldName].args;
         currentQueryArguments.forEach((arg: any) => {
-            if(this.isArgNameOfEntityFilterInputType(arg)) {
+            if (this.isArgNameOfEntityFilterInputType(arg)) {
                 argName = arg.name;
             }
         });
@@ -60,7 +66,10 @@ export class DatesFilterMiddleware {
     }
 
     private isArgNameOfEntityFilterInputType(arg: any) {
-        return arg.type.name === entityFilterInputTypeName || arg.type.ofType?.name === entityFilterInputTypeName;
+        return (
+            arg.type.name === entityFilterInputTypeName ||
+            arg.type.ofType?.name === entityFilterInputTypeName
+        );
     }
 
     private isCurrentFieldASchemaQuery(info: GraphQLResolveInfo): boolean {
@@ -69,19 +78,19 @@ export class DatesFilterMiddleware {
 
     private getEntityDateRangeFilter(args: any): EntityFilter {
         this.formatDatesToIsoString(args);
-        const entityFilter: EntityFilter = {creationTimeFilter: {}, lastUpdateTimeFilter: {}};
-        entityFilter.creationTimeFilter = {...args.creationTime};
-        entityFilter.lastUpdateTimeFilter = {...args.lastUpdateTime};
+        const entityFilter: EntityFilter = { creationTimeFilter: {}, lastUpdateTimeFilter: {} };
+        entityFilter.creationTimeFilter = { ...args.creationTime };
+        entityFilter.lastUpdateTimeFilter = { ...args.lastUpdateTime };
         return entityFilter;
     }
 
     private formatDatesToIsoString(args: any) {
-        Object.keys(args).forEach(key => {
-            Object.keys(args[key]).forEach(val => {
-                if(val) {
+        Object.keys(args).forEach((key) => {
+            Object.keys(args[key]).forEach((val) => {
+                if (val) {
                     args[key][val] = new Date(args[key][val]).toISOString();
                 }
-            })
+            });
         });
     }
 }