+++ packages/polaris-middlewares/src/utills/query-util.ts
@@ -1,3 +1,0 @@
-export function isMutation(query?: string): boolean {
-    return query === undefined ? false : query.trim().startsWith('mutation');
-}