+++ packages/polaris-typeorm/src/handlers/data-version-handler.ts
@@ -1,19 +1,18 @@
 import { PolarisExtensions, PolarisGraphQLContext } from '@enigmatis/polaris-common';
 import { EntityMetadata } from 'typeorm';
 import { RelationMetadata } from 'typeorm/metadata/RelationMetadata';
-import { DataVersion, PolarisConnection, QueryRunner, SelectQueryBuilder } from '..';
+import { DataVersion, PolarisConnection, PolarisEntityManager, SelectQueryBuilder } from '..';
 import { isDescendentOfCommonModel } from '../utils/descendent-of-common-model';
 
 export class DataVersionHandler {
     public async updateDataVersion<Entity>(
-        context: PolarisGraphQLContext,
         connection: PolarisConnection,
-        runner: QueryRunner,
+        manager: PolarisEntityManager,
     ) {
         const extensions: PolarisExtensions =
-            (context && context.returnedExtensions) || ({} as PolarisExtensions);
+            (manager.context && manager.context.returnedExtensions) || ({} as PolarisExtensions);
         connection.logger.log('log', 'Started data version job when inserting/updating entity');
-        const result = await this.getDataVersionForMutation(runner, connection);
+        const result = await this.selectDataVersionForUpdate(manager, connection);
         if (!result) {
             if (extensions.dataVersion) {
                 throw new Error(
@@ -21,52 +20,41 @@ export class DataVersionHandler {
                 );
             }
             connection.logger.log('log', 'no data version found');
-            await runner.manager.save(DataVersion, new DataVersion(1));
+            await manager.save(DataVersion, new DataVersion(1));
             connection.logger.log('log', 'data version created');
             extensions.dataVersion = 2;
         } else {
             if (!extensions.dataVersion) {
                 connection.logger.log('log', 'context does not hold data version');
-                extensions.dataVersion = result.getValue() + 1;
-                await runner.manager.increment(DataVersion, {}, 'value', 1);
+                extensions.dataVersion = result.value + 1;
+                await manager.increment(DataVersion, {}, 'value', 1);
                 connection.logger.log('log', 'data version is incremented and holds new value');
             } else {
-                if (extensions.dataVersion !== result.getValue()) {
+                if (extensions.dataVersion !== result.value) {
                     throw new Error('data version in context does not equal data version in table');
                 }
             }
         }
-        if (context && extensions) {
-            context.returnedExtensions = extensions;
+        if (manager.context && extensions) {
+            manager.context.returnedExtensions = extensions;
         }
-        connection.logger.log(
-            'log',
-            'Finished data version job when inserting/updating entity',
-            runner,
-        );
+        connection.logger.log('log', 'Finished data version job when inserting/updating entity');
     }
-    private async getDataVersionForMutation(
-        runner: any,
-        connection: any,
-    ): Promise<DataVersion | undefined> {
-        let result;
+    private async selectDataVersionForUpdate(
+        manager: PolarisEntityManager,
+        connection: PolarisConnection,
+    ): Promise<{ id: number; value: number } | undefined> {
         try {
-            if (!runner.isTransactionActive) {
-                await runner.startTransaction();
-            }
-            result = await runner.manager
-                .getRepository(DataVersion)
+            return manager
                 .createQueryBuilder()
+                .from(DataVersion, 'dv')
                 .useTransaction(true)
                 .setLock('pessimistic_write')
-                .andWhereInIds(1)
-                .getOne();
+                .getRawOne();
         } catch (e) {
-            connection.logger.log('warn', 'waiting for lock of data version to release');
-            await runner.rollbackTransaction();
-            result = this.getDataVersionForMutation(runner, connection);
+            connection.logger.log('warn', e);
+            throw e;
         }
-        return result;
     }
 }
 
@@ -106,10 +94,10 @@ function getQbWithSelect(
     selectAll: boolean,
 ) {
     return selectAll
-        ? qb.leftJoinAndSelect(`${entityMetadata.tableName}.${relation.propertyName}`, alias)
+        ? qb.leftJoinAndSelect(`${entityMetadata.name}.${relation.propertyName}`, alias)
         : qb
-              .leftJoin(`${entityMetadata.tableName}.${relation.propertyName}`, alias)
-              .addSelect(alias + '.dataVersion');
+              .leftJoin(`${entityMetadata.name}.${relation.propertyName}`, alias)
+              .addSelect(`${alias}.dataVersion`);
 }
 
 function leftJoinRelationsToQB(
@@ -123,7 +111,7 @@ function leftJoinRelationsToQB(
     const childDVMapping = getPropertyMap(children, relation);
     if (childDVMapping) {
         const relationMetadata = relation.inverseEntityMetadata;
-        const alias: string = relationMetadata.tableName;
+        const alias: string = relationMetadata.name;
         const notInJoins = names.filter((x) => x === alias).length === 0;
         if (isDescendentOfCommonModel(relationMetadata) && notInJoins) {
             qb = getQbWithSelect(qb, entityMetadata, relation, alias, selectAll);
@@ -168,16 +156,27 @@ function joinDataVersionRelations(
     entityMetadata: EntityMetadata,
     names: string[],
 ): SelectQueryBuilder<any> {
+    let entityIdAlias = entityMetadata.name;
+    if (findSorted) {
+        qb.select(`${entityIdAlias}.id`, entityIdAlias);
+    }
     if (context.dataVersionContext?.mapping && shouldLoadRelations) {
-        if (findSorted) {
-            qb.select(entityName + '.id', 'id').addSelect(entityName + '.dataVersion');
-        }
         qb = loadRelations(
             qb,
             entityMetadata,
             names,
             context.dataVersionContext!.mapping!,
             !findSorted,
+        );
+    }
+    if (qb.expressionMap.selects.length <= 1 && findSorted) {
+        entityIdAlias = 'entityId';
+        qb.select(`${entityName}.id`, entityIdAlias);
+    }
+    if (findSorted) {
+        qb.addSelect(
+            `${entityIdAlias}.dataVersion`,
+            qb.expressionMap.selects.length > 1 ? undefined : 'maxDV',
         );
     }
     return qb;